Changed returned id to array index
This abstracts away the process of directly reading / writing
preferences, allowing the plugin to cache values instead of having to
query all reads.
It also opens up the possibility of finer control of what plugin can
modify which category.

The downside is that the number of categories is limited by a hardcoded
value in the plugin.

To compensate, functions to directly access the database are provided.