Added preference caching
I'm not quite sure if there really is a benefit to allowing caching at
the moment, though it should take a load off of having to read the
preference from the database every time.  It does use a bit more memory,
though.

There may need to be some callback work involved so queries can be
threaded.